@@ -18,21 +18,22 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.1+5
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: ^3.5.4
 
 dependencies:
   flutter:
     sdk: flutter
   provider: ^6.1.2
   shared_preferences: ^2.3.5
-  app_review: ^2.1.2+1
-
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
   audioplayers: ^6.1.0
   flutter_launcher_icons: ^0.14.2
+  # app_review: ^2.1.2+1
+  in_app_review: ^2.0.9
+  # app_review: ^2.1.2+1
 
 dev_dependencies:
   flutter_test:
